+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Support Instance Warmup via Annotation.
+
 ## [0.2.0] - 2022-09-20
 
 ### Changed